+++ src/include/rest_catalog/objects/view_metadata.hpp
@@ -31,9 +31,9 @@ public:
 
 public:
 	string view_uuid;
-	int64_t format_version;
+	int32_t format_version;
 	string location;
-	int64_t current_version_id;
+	int32_t current_version_id;
 	vector<ViewVersion> versions;
 	vector<ViewHistoryEntry> version_log;
 	vector<Schema> schemas;
